revised consolelogin logic
